requirements.txt
pandas==1.1.3
requests==2.24.0
selenium==3.141.0
attrs~=20.3.0